validation: crypto: move code to separate files
Move test_crypto_op() and its implementation functions in a separate
file to improve readability. Move test vector definitions to a separate
file to make it possible to include test_vector.h in more than one
compilation unit without warnings about unused definitions. Move also
some common code from the main code file to a separate file.
